log/requirements.txt
beautifulsoup4==4.10.0
comtypes==1.1.10
python-docx==0.8.11
requests==2.25.1
selenium==4.0.0
webdriver-manager==3.5.2

